Update Broadlink LAN SmartApp
This smartapp requires that you add settings as described in the readme, then install at the location where the smart things hub is.  This app imports your RF and IR device codes automatically but does not poll for changes.  When new devices are added or changed, just install again and it will re-run the import.