Prevent re-sending of commands
Currently, the device handler will send the IR codes to the RM even if the same code has already been issued.  For most devices this may not be a problem, but for devices that respond with confirmation of the code (i.e. an air conditioner beep), this can get annoying if coupled with a smart app that keeps sending "on" commands.

The code above stores the "on/off" state of the device in the smartthings state store and checks if it is in the current state or not before issuing the IR code.